refactor(execution): rename workflow catalog surface
